Creating a separate benchmark gradle task
Moving the benchmark classes to the main source set, and creating a
separate benchmark task that actually runs the benchmarks with the full
duration.

Benchmarks can still be tested using a minimum configuration with
TestRunners.minimalRunner(), which will automatically tune the test down
the smallest runtime. Tests for individual benchmarks should be added to
the test source set.