GEODE-6081: Probe that uses HdrHistogram
Adding a new probe that uses HdrHistogram to generate a summary with a
mean and 99th percentile latency.

HdrHistogram should be more accurate than yardsticks histogram probe.

This probe currently just reports a one line summary, but we could
consider replacing the existing percentile and throughtput probes with
this thing in the future - it has its own file format for recording per
sample histograms.