GEODE-6086: Handling multiple clients in HdrHistogramProbe
This probe was generating a single number in the csv file, but to correctly
combine results from multiple clients we need to output the entire histogram
for each client and combine them before getting the 99th percentile latency.

Writing the histogram results to a new .hlog file and combining them when
parsing the results.