GEODE-6145: Allow multiple results for a single probe analyzer (#14)
This change modifies the ProbeResultParser interface to return a list of
ResultData objects.  It takes advantage of this new functionality to return
6 results from the HDRHistogramParser (median latency, 90, 99, 99.9 percentile
latencies, average latency, and latency standard deviation).  This is probably
too noisy for the long run, but could prove useful as we evaluate different
platforms.

This change also changes the BenchmarkRunAnalyzer (invoked via gradlew
analyzeRun) to take the baseline directory as its first argument.