GEODE-6148: improve performance of Prepopulate
Prepopulating the region used to occur by means of puts from every
server for every key in the specified range. This resulted in
unnecessarily conflicting puts.

To improve performance of the setup, PrePopulateRegion.run() now divides
the key range as evenly as possible between all servers in the cluster,
prepares the data in a map, then puts the map to the region using
putAll() in threads.

Also:
* add redundancy to Partitioned Regions
* increase JVM Memory

The changes result in an observable improvement in performance.

Signed-off-by: Dan Smith <dsmith@pivotal.io>