Update script for running tests on AWS
Update the script for running tests on AWS to use the default version
number. The gradle parameters for specifying a custom version number
change in different versions. To have one script that works for all
versions, we use the default version number then get that version from
gfsh for use in benchmarks.

Co-Authored-By: jbarrett@pivotal.io